+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatComboBoxUI.java
@@ -61,6 +61,7 @@ import javax.swing.UIManager;
 import javax.swing.border.AbstractBorder;
 import javax.swing.border.Border;
 import javax.swing.plaf.ComponentUI;
+import javax.swing.plaf.UIResource;
 import javax.swing.plaf.basic.BasicComboBoxUI;
 import javax.swing.plaf.basic.BasicComboPopup;
 import javax.swing.plaf.basic.ComboPopup;
@@ -97,11 +98,12 @@ import com.formdev.flatlaf.util.UIScale;
  * @uiDefault ComboBox.disabledForeground		Color
  * @uiDefault ComboBox.buttonBackground			Color
  * @uiDefault ComboBox.buttonEditableBackground	Color
- * @uiDefault ComboBox.buttonFocusedBackground	Color	optional
+ * @uiDefault ComboBox.buttonFocusedBackground	Color	optional; defaults to ComboBox.focusedBackground
  * @uiDefault ComboBox.buttonArrowColor			Color
  * @uiDefault ComboBox.buttonDisabledArrowColor	Color
  * @uiDefault ComboBox.buttonHoverArrowColor	Color
  * @uiDefault ComboBox.buttonPressedArrowColor	Color
+ * @uiDefault ComboBox.popupFocusedBackground	Color	optional
  *
  * @author Karl Tauber
  */
@@ -117,8 +119,8 @@ public class FlatComboBoxUI
 	protected Color disabledBorderColor;
 
 	protected Color editableBackground;
-	protected Color disabledBackground;
 	protected Color focusedBackground;
+	protected Color disabledBackground;
 	protected Color disabledForeground;
 
 	protected Color buttonBackground;
@@ -128,6 +130,8 @@ public class FlatComboBoxUI
 	protected Color buttonDisabledArrowColor;
 	protected Color buttonHoverArrowColor;
 	protected Color buttonPressedArrowColor;
+
+	protected Color popupFocusedBackground;
 
 	private MouseListener hoverListener;
 	protected boolean hover;
@@ -199,8 +203,8 @@ public class FlatComboBoxUI
 		disabledBorderColor = UIManager.getColor( "Component.disabledBorderColor" );
 
 		editableBackground = UIManager.getColor( "ComboBox.editableBackground" );
-		disabledBackground = UIManager.getColor( "ComboBox.disabledBackground" );
 		focusedBackground = UIManager.getColor( "ComboBox.focusedBackground" );
+		disabledBackground = UIManager.getColor( "ComboBox.disabledBackground" );
 		disabledForeground = UIManager.getColor( "ComboBox.disabledForeground" );
 
 		buttonBackground = UIManager.getColor( "ComboBox.buttonBackground" );
@@ -210,6 +214,8 @@ public class FlatComboBoxUI
 		buttonDisabledArrowColor = UIManager.getColor( "ComboBox.buttonDisabledArrowColor" );
 		buttonHoverArrowColor = UIManager.getColor( "ComboBox.buttonHoverArrowColor" );
 		buttonPressedArrowColor = UIManager.getColor( "ComboBox.buttonPressedArrowColor" );
+
+		popupFocusedBackground = UIManager.getColor( "ComboBox.popupFocusedBackground" );
 
 		// set maximumRowCount
 		int maximumRowCount = UIManager.getInt( "ComboBox.maximumRowCount" );
@@ -230,8 +236,8 @@ public class FlatComboBoxUI
 		disabledBorderColor = null;
 
 		editableBackground = null;
-		disabledBackground = null;
 		focusedBackground = null;
+		disabledBackground = null;
 		disabledForeground = null;
 
 		buttonBackground = null;
@@ -241,6 +247,8 @@ public class FlatComboBoxUI
 		buttonDisabledArrowColor = null;
 		buttonHoverArrowColor = null;
 		buttonPressedArrowColor = null;
+
+		popupFocusedBackground = null;
 
 		MigLayoutVisualPadding.uninstall( comboBox );
 	}
@@ -431,7 +439,11 @@ public class FlatComboBoxUI
 
 			// paint arrow button background
 			if( enabled && !isCellRenderer ) {
-				g2.setColor( paintButton ? buttonEditableBackground : buttonFocusedBackground != null && isFocusOwner() ? buttonFocusedBackground : buttonBackground );
+				g2.setColor( paintButton
+					? buttonEditableBackground
+					: (buttonFocusedBackground != null || focusedBackground != null) && isPermanentFocusOwner( comboBox )
+						? (buttonFocusedBackground != null ? buttonFocusedBackground : focusedBackground)
+						: buttonBackground );
 				Shape oldClip = g2.getClip();
 				if( isLeftToRight )
 					g2.clipRect( arrowX, 0, width - arrowX, height );
@@ -491,15 +503,20 @@ public class FlatComboBoxUI
 	}
 
 	protected Color getBackground( boolean enabled ) {
-		return enabled
-			? (focusedBackground != null && isFocusOwner()
-				? focusedBackground
-				: (editableBackground != null && comboBox.isEditable() ? editableBackground : comboBox.getBackground()) )
-			: (isIntelliJTheme ? FlatUIUtils.getParentBackground( comboBox ) : disabledBackground);
-	}
-	
-	protected boolean isFocusOwner() {
-		return FlatUIUtils.isPermanentFocusOwner( comboBox ) || comboBox.getEditor() != null && comboBox.getEditor().getEditorComponent() != null && FlatUIUtils.isPermanentFocusOwner( comboBox.getEditor().getEditorComponent() );
+		if( enabled ) {
+			Color background = comboBox.getBackground();
+
+			// always use explicitly set color
+			if( !(background instanceof UIResource) )
+				return background;
+
+			// focused
+			if( focusedBackground != null && isPermanentFocusOwner( comboBox ) )
+				return focusedBackground;
+
+			return (editableBackground != null && comboBox.isEditable()) ? editableBackground : background;
+		} else
+			return isIntelliJTheme ? FlatUIUtils.getParentBackground( comboBox ) : disabledBackground;
 	}
 
 	protected Color getForeground( boolean enabled ) {
@@ -710,9 +727,8 @@ public class FlatComboBoxUI
 			super.configureList();
 
 			list.setCellRenderer( new PopupListCellRenderer() );
-			if( focusedBackground != null ) {
-		        list.setBackground( focusedBackground );
-			}
+			if( popupFocusedBackground != null )
+		        list.setBackground( popupFocusedBackground );
 		}
 
 		@Override